trading_bot/requirements.txt
aiogram>=3.5.0
aiosqlite>=0.20.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
uvloop; platform_system == "Linux"

